AP_RobotisServo: more changes for 32 bit servo mask

+++ libraries/AP_RobotisServo/AP_RobotisServo.h
@@ -58,7 +58,7 @@ private:
     void configure_servos(void);
 
     // auto-detected mask of available servos, from a broadcast ping
-    uint16_t servo_mask;
+    uint32_t servo_mask;
     uint8_t detection_count;
     uint8_t configured_servos;
     bool initialised;
